requirements.txt
aiogram>=3.10.0
aportalsmp>=0.1.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
uvloop>=0.20.0; platform_system != "Windows"
